INTEXT-209: AMQP: Fix container registration
JIRA: https://jira.spring.io/browse/INTEXT-209

An internal `SimpleMessageListenerContainer` instance from the `Amqp.inboundGateway()/inboundAdapter()`
hasn't been registered as a bean in the application context. That causes issues when container interacts with the applicationContext.

Add `implements ComponentsRegistration` to the `AmqpInboundGatewaySpec` and `AmqpInboundChannelAdapterSpec`.
Add simple `stop()/start()` operation to the `AmqpTests` to be sure that the issue has gone.

**Cherry-pick to 1.0.x**