INTEXT-217: Cleanup KafkaProducerContext.beanName
JIRA: https://jira.spring.io/browse/INTEXT-217

The `KafkaProducerMessageHandlerSpec` provides an internal `KafkaProducerContext` for the bean registration.
The `KafkaProducerContext` is `NamedComponent` and we don't generate `beanName` for it because its value is `not_specified` by default.
That causes the `duplicate bean` issue in case of several ` Kafka.outboundChannelAdapter()`

* Clean the `KafkaProducerContext.beanName` in the `KafkaProducerMessageHandlerSpec` to allow `IntegrationFlowBeanPostProcessor` to generate
an unique `beanName` for each `KafkaProducerContext` bean definition.