Auto-configure JMX to ensure MBeanServer bean
Spring Boot 1.3.4 will fix a bug where an MBeanServer bean was being
auto-configured even though auto-configuration of JMX had been
disabled due to some faulty logic in IntegrationAutoConfiguration
(see spring-projects/spring-boot#5309). A side-effect of this is that
FtpTests, JmsTests and SftpTests fail due to a missing mbeanServer
bean.

This commit updates the failing tests to import JmxAutoConfiguration,
thereby ensuring that an mbeanServer bean is auto-configured.