Optimize `Tuple`s
Previously the `Tuple` implementation has been based on the `Object[] entries`
representation for the values.
That is a bit suboptimal from the memory allocating perspective.
Plus that add some micro overhead during index accessing for items.

Change the `Tuple1` and `Tuple2` to be as simple POJO with properties for better memory
consumption and runtime usage.

* Adjust to the latest Spring Kafka