GH-69: Add `MessageChannelSpec.wireTap()`
Fixes GH-69 (https://github.com/spring-projects/spring-integration-java-dsl/issues/69)

* Add explicit `wireTap` declaration to the `MessageChannelSpec`.
That's allow to simplify the configuration for short use-case bypassing the `MessageChannel` and `WireTap` bean definition.
* For this purpose allow from the Framework perspective to declare `IntegrationComponentSpec` as a `@Bean`
and convert it to the target bean using its `.get()` from the `IntegrationFlowBeanPostProcessor`
* In addition extract the Spec's `getComponentsToRegister()` in the `IntegrationFlowBeanPostProcessor` and register them as beans, too.
That is useful exactly for that `.wireTap()` from the `MessageChannelSpec`, when we should to register `WireTap` and its `selector` as beans anyway.
* Make `WireTapSpec` as public class according end-users experience to have more cleaner and shorter configs
* Make `IntegrationFlowDefinition.wireTap(WireTapSpec wireTapSpec)` as `public` for convenience.