GH-68: Allow any type mapping for routers
Fixes GH-68 (https://github.com/spring-projects/spring-integration-java-dsl/issues/68)

The Spring Integration Core allows to map channels in the router only via `String` key.
With Java configuration it isn't so convenient when we would like to map and resolve to any type.

That's also lets Java DSL to resolve generics at typing time, when we provide a routing `Function`,
so `channelMapping` and `subFlowMapping` coerce to the result type if any.