GH-104: Disallow Scope for `IntegrationFlow`
Fixes GH-104 (https://github.com/spring-projects/spring-integration-java-dsl/issues/104)

Since `IntegrationFlow` is just a logical container for the rest Integration component in the flow definition,
it doesn't not make any sense to have it in the `scope`. Just because all those components are registered as singletons for their proper stateless lifecycle and logic

* Throw `BeanCreationNotAllowedException` if `IntegrationFlow` bean definition is in not singleton scope.

* Fix Scope detection using `getBeanDefinitionNames()` instead of real `getBean()`
* Optimize `IntegrationFlow` process in the `IntegrationFlowBeanPostProcessor` via real object replacing its `integrationComponents`.
That allow do not jump to the `beanFactory` from the `IntegrationFlowContext`
* Don't register `IntegrationFlow` as a bean in the `IntegrationFlowContext`

Fixes after rebase.

Leave the manual flow registration as a bean though.

* Avoid `StandardIntegrationFlow` mutation via new `setIntegrationComponents()`
* Redo `Lifecycle` logic in the `StandardIntegrationFlow`

Fix condition when cancelled task in the taskScheduler might not be cancelled,
because of `receiveTimeout` race condition on the `QueueChannel.receive()`.

* Add `.receiveTimeout(10)` to the testing flow.
* Also add `.maxMessagesPerPoll(1)` to exit from polling task as fast as possible